+++ lib/api/lib/Paymill/API/Curl.php
@@ -41,7 +41,7 @@ class Curl extends CommunicationAbstract
      *   Extra cURL options. The array is keyed by the name of the cURL
      *   options.
      */
-    public function __construct($apiKey, $apiEndpoint = 'https://api.paymill.com/v2/', array $extracURL = array())
+    public function __construct($apiKey, $apiEndpoint = 'https://api.paymill.com/v2.1/', array $extracURL = array())
     {
         $this->_apiKey = $apiKey;
         $this->_apiUrl = $apiEndpoint;
@@ -67,8 +67,7 @@ class Curl extends CommunicationAbstract
             CURLOPT_RETURNTRANSFER => true,
             CURLOPT_CUSTOMREQUEST => $method,
             CURLOPT_USERAGENT => 'Paymill-php/0.0.2',
-            CURLOPT_SSL_VERIFYPEER => true,
-            CURLOPT_CAINFO => realpath(dirname(__FILE__)) . DIRECTORY_SEPARATOR . 'paymill.crt',
+            CURLOPT_SSL_VERIFYPEER => true
         );
 
         // Add extra options to cURL if defined.
@@ -88,7 +87,6 @@ class Curl extends CommunicationAbstract
         if ($this->_apiKey) {
             $curlOpts[CURLOPT_USERPWD] = $this->_apiKey . ':';
         }
-
         $curl = curl_init();
         curl_setopt_array($curl, $curlOpts);
         $responseBody = $this->_curlExec($curl);
@@ -101,6 +99,10 @@ class Curl extends CommunicationAbstract
 
         if ('application/json' === $responseInfo['content_type']) {
             $responseBody = json_decode($responseBody, true);
+        } elseif('text/csv' === $responseInfo['content_type']
+            && !isset($responseBody['error'])
+        ) {
+            return $responseBody;
         }
 
         return array(