+++ lib/api/lib/Paymill/Models/Request/Preauthorization.php
@@ -10,26 +10,35 @@ namespace Paymill\Models\Request;
  */
 class Preauthorization extends Base
 {
-
     /**
      * @var string
      */
     private $_amount;
-    
+
     /**
      * @var string
      */
     private $_currency;
-    
+
     /**
      * @var string
      */
     private $_payment;
-    
+
     /**
      * @var string
      */
     private $_token;
+
+    /**
+     * @var string
+     */
+    private $_description;
+
+    /**
+     * @var string
+     */
+    private $_client;
 
     /**
      * Creates an instance of the preauthorization request model
@@ -120,6 +129,46 @@ class Preauthorization extends Base
     }
 
     /**
+     * Returns the description
+     * @return string
+     */
+    public function getDescription()
+    {
+        return $this->_description;
+    }
+
+    /**
+     * Sets the description
+     * @param string $description
+     * @return \Paymill\Models\Request\Preauthorization
+     */
+    public function setDescription($description)
+    {
+        $this->_description = $description;
+        return $this;
+    }
+
+    /**
+     * Returns the client
+     * @return string
+     */
+    public function getClient()
+    {
+        return $this->_client;
+    }
+
+    /**
+     * Sets the client
+     * @param string $client
+     * @return \Paymill\Models\Request\Preauthorization
+     */
+    public function setClient($client)
+    {
+        $this->_client = $client;
+        return $this;
+    }
+
+    /**
      * Returns an array of parameters customized for the argumented methodname
      * @param string $method
      * @return array
@@ -136,6 +185,10 @@ class Preauthorization extends Base
                 }
                 $parameterArray['amount'] = $this->getAmount();
                 $parameterArray['currency'] = $this->getCurrency();
+                $parameterArray['description'] = $this->getDescription();
+                if (!is_null($this->getClient())) {
+                    $parameterArray['client'] = $this->getClient();
+                }
 
                 break;
             case 'getOne':
@@ -143,7 +196,7 @@ class Preauthorization extends Base
                 $parameterArray['offset'] = 0;
                 break;
             case 'getAll':
-            $parameterArray = $this->getFilter();
+                $parameterArray = $this->getFilter();
                 break;
             case 'delete':
                 break;