+++ lib/api/tests/integration/ClientTest.php
@@ -28,7 +28,10 @@ class Client extends PHPUnit_Framework_TestCase
     protected function setUp()
     {
         $this->_service = new Request();
-        $this->_service->setConnectionClass(new Curl(API_TEST_KEY));
+        $this->_service->setConnectionClass(
+            new Curl(API_TEST_KEY, API_HOST, array(CURLOPT_SSL_VERIFYPEER => SSL_VERIFY_PEER))
+        );
+
         $this->_model = new Models\Request\Client();
         parent::setUp();
     }
@@ -71,6 +74,7 @@ class Client extends PHPUnit_Framework_TestCase
     }
 
     /**
+     *
      * @test
      * @codeCoverageIgnore
      * @expectedException \Paymill\Services\PaymillException
@@ -110,16 +114,30 @@ class Client extends PHPUnit_Framework_TestCase
     /**
      * @test
      * @codeCoverageIgnore
+     * @depends createClient
+     */
+    public function getAllClientAsModel()
+    {
+        $result = $this->_service->getAllAsModel($this->_model);
+        $this->assertInternalType('array', $result, var_export($result, true));
+        $this->assertInstanceOf('Paymill\Models\Response\Client', array_pop($result));
+    }
+
+
+
+    /**
+     * @test
+     * @codeCoverageIgnore
      */
     public function getAllClientWithFilter()
     {
         $this->_model->setFilter(array(
-            'count' => 2,
+            'count' => 1,
             'offset' => 0
             )
         );
         $result = $this->_service->getAll($this->_model);
-        $this->assertEquals(2, count($result), var_export($result, true));
+        $this->assertEquals(1, count($result), var_export($result, true));
     }
 
     /**