+++ lib/api/tests/integration/PreauthorizationTest.php
@@ -28,7 +28,10 @@ class PreauthorizationTest extends PHPUnit_Framework_TestCase
     protected function setUp()
     {
         $this->_service = new Request();
-        $this->_service->setConnectionClass(new Curl(API_TEST_KEY));
+        $this->_service->setConnectionClass(
+            new Curl(API_TEST_KEY, API_HOST, array(CURLOPT_SSL_VERIFYPEER => SSL_VERIFY_PEER))
+        );
+
         $this->_model = new Models\Request\Preauthorization();
         parent::setUp();
     }
@@ -54,6 +57,7 @@ class PreauthorizationTest extends PHPUnit_Framework_TestCase
             ->setCurrency('EUR');
         $result = $this->_service->create($this->_model);
         $this->assertInstanceOf('Paymill\Models\Response\Preauthorization', $result);
+        $this->assertNotNull($result.getTransaction());
         return $result;
     }
 
@@ -102,16 +106,29 @@ class PreauthorizationTest extends PHPUnit_Framework_TestCase
     /**
      * @test
      * @codeCoverageIgnore
+     * @depends createPreauthorization
+     */
+    public function getAllPreauthorizationAsModel()
+    {
+        $this->_model;
+        $result = $this->_service->getAllAsModel($this->_model);
+        $this->assertInternalType('array', $result, var_export($result, true));
+		$this->assertInstanceOf('Paymill\Models\Response\Preauthorization', array_pop($result));
+    }
+
+    /**
+     * @test
+     * @codeCoverageIgnore
      */
     public function getAllPreauthorizationWithFilter()
     {
         $this->_model->setFilter(array(
-            'count' => 2,
+            'count' => 1,
             'offset' => 0
             )
         );
         $result = $this->_service->getAll($this->_model);
-        $this->assertEquals(2, count($result), var_export($result, true));
+        $this->assertEquals(1, count($result), var_export($result, true));
     }
 
     /**