+++ lib/api/tests/integration/WebhookTest.php
@@ -28,7 +28,9 @@ class WebhookTest extends PHPUnit_Framework_TestCase
     protected function setUp()
     {
         $this->_service = new Request();
-        $this->_service->setConnectionClass(new Curl(API_TEST_KEY));
+        $this->_service->setConnectionClass(
+            new Curl(API_TEST_KEY, API_HOST, array(CURLOPT_SSL_VERIFYPEER => SSL_VERIFY_PEER))
+        );
         $this->_model = new Models\Request\Webhook();
         $this->_email       = 'dummy@example.com';
         $this->_url         = 'http://example.com/dummyCallback';
@@ -52,11 +54,13 @@ class WebhookTest extends PHPUnit_Framework_TestCase
     public function createWebhookWithUrl()
     {
         $this->_model->setUrl('http://example.com/dummyCallback')
+            ->setActive(true)
             ->setEventTypes(array(
                 'transaction.succeeded', 'subscription.created'
             ));
         $result = $this->_service->create($this->_model);
         $this->assertInstanceOf('Paymill\Models\Response\Webhook', $result, var_export($result, true));
+        $this->assertTrue($result->getActive());
         return $result;
     }
 
@@ -67,6 +71,7 @@ class WebhookTest extends PHPUnit_Framework_TestCase
     public function createWebhookWithEmail()
     {
         $this->_model->setEmail('dummy@example.com')
+            ->setActive(true)
             ->setEventTypes(array(
                 'transaction.succeeded', 'subscription.created'
             ));
@@ -83,11 +88,13 @@ class WebhookTest extends PHPUnit_Framework_TestCase
     public function updateWebhook($model)
     {
         $this->_model->setId($model->getId())
+            ->setActive(false)
             ->setUrl('http://example.com/dummyCallbackUpdate');
         $result = $this->_service->update($this->_model);
 
         $this->assertInstanceOf('Paymill\Models\Response\Webhook', $result, var_export($result, true));
         $this->assertEquals($model->getId(), $result->getId());
+        $this->assertFalse($result->getActive());
     }
 
     /**
@@ -117,16 +124,30 @@ class WebhookTest extends PHPUnit_Framework_TestCase
     /**
      * @test
      * @codeCoverageIgnore
+     * @depends createWebhookWithUrl
+     */
+    public function getAllWebhookAsModel()
+    {
+        $this->_model;
+        $result = $this->_service->getAllAsModel($this->_model);
+        $this->assertInternalType('array', $result, var_export($result, true));
+		$this->assertInstanceOf('Paymill\Models\Response\Webhook', array_pop($result));
+    }
+
+    /**
+     * @test
+     * @depends createWebhookWithUrl
+     * @codeCoverageIgnore
      */
     public function getAllWebhookWithFilter()
     {
         $this->_model->setFilter(array(
-            'count' => 2,
+            'count' => 1,
             'offset' => 0
             )
         );
         $result = $this->_service->getAll($this->_model);
-        $this->assertEquals(2, count($result), var_export($result, true));
+        $this->assertEquals(1, count($result), var_export($result, true));
     }
 
     /**