+++ paymill.php
@@ -3,15 +3,15 @@
 Plugin Name: Paymill
 Plugin URI: https://www.paymill.com
 Description: Payments made easy
-Version: 1.8.2
+Version: 1.9
 Author: Matthias Reuter info@straightvisions.com
 Author URI: http://elbnetz.com
 */
 
 	// common information
-	define('PAYMILL_VERSION',1802);
+	define('PAYMILL_VERSION',1900);
 	define('PAYMILL_DIR',WP_PLUGIN_DIR.'/'.dirname(plugin_basename(__FILE__)).'/');
-	define('PAYMILL_PLUGIN_URL',plugins_url( '' , __FILE__ ).'/');
+	define('PAYMILL_PLUGIN_URL',str_replace('http:','',plugins_url( '' , __FILE__ ).'/'));
 	$GLOBALS['paymill_active'] = false; // eCommerce channels will set Paymill as active later to prevent showing payment form twice on same page.
 
 	// service mode